test: refactor helper functions out of regex_test.js
